// data/github/current/ui/packages/copilot-for-business/types.ts
export const CopilotForBusinessSeatPolicy = {
  Disabled: 'disabled',
  EnabledForAll: 'enabled_for_all',
  EnabledForSelected: 'enabled_for_selected',
} as const

export type CopilotForBusinessSeatPolicy =
  (typeof CopilotForBusinessSeatPolicy)[keyof typeof CopilotForBusinessSeatPolicy]

export const SeatType = {
  OrganizationInvitation: 'OrganizationInvitation',
  Team: 'Team',
  User: 'User',
  Organization: 'Organization',
  All: 'All',
  EnterpriseTeam: 'EnterpriseTeam',
} as const

export type SeatType = (typeof SeatType)[keyof typeof SeatType]

export type SeatBreakdown = {
  seats_assigned: number
  seats_billed: number
  seats_pending: number
  description: string
}

export type CopilotForBusinessTrial = {
  started: boolean
  ended: boolean
  has_trial: boolean
  upgradable: boolean
  cancelable: boolean
  days_left: number
  started_at: string
  ends_at: string
  trial_length: number
  active: boolean
  expired: boolean
  pending: boolean
  copilot_plan: string
  belongs_to_trial_business_account: boolean
}

export type CopilotForBusinessPayload = {
  policy: CopilotForBusinessSeatPolicy
  public_code_suggestions_configured: boolean
  seats: {
    seats: CopilotSeatAssignment[]
    count: number
    pending_requests?: {
      requesters: Requester[]
      count: number
    }
    licenses: CopilotLicenseIdentifiers
  }
  organization: CopilotForBusinessOrganization
  business?: CopilotForBusinessBusiness
  seat_breakdown: SeatBreakdown
  seat_assignments: number
  business_trial?: CopilotForBusinessTrial
  render_trial_expired_banner: boolean
  render_copilot_insights_banner: boolean
  can_add_teams: boolean
  members_count: number
  can_allow_to_assign_seats_on_business: boolean
  next_billing_date: string
  plan_text: 'Business' | 'Enterprise'
  featureRequestInfo: FeatureRequestInfo
  render_pending_downgrade_banner: boolean
  display_activity: boolean
  adoption_metrics: {
    total: number
    active: number
    inactive: number
    dormant: number
  }
}

export type CopilotLicenseIdentifiers = {
  user_ids: number[]
  team_ids: number[]
  invite_user_ids: number[]
  invite_emails: string[]
}

export type CopilotForBusinessOrganization = {
  name: string
  id: number
  billable: boolean
  has_seat: boolean
  add_seat_link: string | null
  copilot_enabled: boolean
}

export type CopilotForBusinessBusiness = {
  name: string
  slug: string
}

export type CopilotForBusinessTeam = {
  id?: number
  login: string
  slug: string
  avatar_url: string
  combined_slug: string
  member_count: number
  member_ids: number[]
}

export type CopilotEnterpriseTeam = {
  id: number
  login: string
  slug: string
  mapping_id?: number // this points to the external group mapping id
  member_count: number
  member_ids: number[]
}

export type CopilotSeatAssignable = {
  id: number
  login?: string
  avatar_url?: string
  display_name?: string
  slug?: string
  combined_slug?: string
  member_count?: number
  member_ids?: number[]
  email?: string
  invitee?: CopilotSeatAssignable
  invitee_id?: number
  mapping_id?: number
}

export type CopilotSeatAssignment<
  Type extends SeatType = SeatType,
  Assignable extends CopilotSeatAssignable = CopilotSeatAssignable,
> = {
  assignable: Assignable
  assignable_type: Type
  id?: number | null
  invitation_date?: string
  invitation_expired?: boolean
  last_activity_at?: string
  pending_cancellation_date?: string | null
  access_revoked_at?: string | null
}
// ...
export type CSVUser = {
  id: number
  email: string
  display_login: string
  profile_name: string
  avatar: string
  is_new_user: boolean
  email_input_csv: boolean
}

export type CSVUploadPayload = {
  // Number of users outside of the org
  new_users: number
  // Users retrieved from the DB via their display login
  github_users: CSVUser[]
  // Users retrieved from the DB via their email address
  email_users: string[]
  // The number of invalid user logins or emails
  found_errors: string[]
  // Total number of users
  total_users: number
}

export type AutocompleteItem = {text: string; id: string}

export type MenuItem = {
  id: string
  selected: boolean
  title: string
  description: string
  value: string
}

type CopilotForBusinessPolicy = {
  manages: string
  configurable: boolean
  options: MenuItem[]
  visible: boolean | null
}

export type CopilotForBusinessGeneralPolicy = {
  manages: string
  configurable: boolean
  options: MenuItem[]
  visible: boolean | null
  helpurl: string
  helptext: string
  displayname: string
  description: string
  preview: boolean
}

type BingGitHubChatPolicy = {
  manages: string
  configurable: boolean
  enabled: boolean
  options: MenuItem[]
  visible: boolean | null
}

type UserFeedbackPolicy = {
  manages: string
  configurable: boolean
  visible: boolean | null
  enabled: boolean
}

type BetaFeaturesPolicy = {
  manages: string
  configurable: boolean
  visible: boolean | null
  enabled: boolean
}

export type CopilotForBusinessPoliciesPayload = {
  org_name: string
  copilot_plan: string
  editor_chat: CopilotForBusinessPolicy
  mobile_chat: CopilotForBusinessPolicy
  snippy: CopilotForBusinessPolicy
  cli: CopilotForBusinessPolicy
  desktop: CopilotForBusinessPolicy
  copilot_for_dotcom: CopilotForBusinessPolicy
  bing_github_chat: BingGitHubChatPolicy
  copilot_user_feedback_opt_in: UserFeedbackPolicy
  copilot_beta_features_opt_in: BetaFeaturesPolicy
  copilot_extensions: CopilotForBusinessPolicy
  private_telemetry: CopilotForBusinessPolicy
  copilot_usage_metrics_policy: CopilotForBusinessPolicy
  editor_preview_features: CopilotForBusinessPolicy
  a_chat: CopilotForBusinessPolicy
  a_chat_ga: boolean
  a_f: CopilotForBusinessPolicy | CopilotForBusinessGeneralPolicy
  afos: CopilotForBusinessGeneralPolicy
  al: CopilotForBusinessGeneralPolicy
  a_f_ga: boolean
  g_chat: CopilotForBusinessPolicy
  g_chat_ga: boolean
  g_tf: CopilotForBusinessGeneralPolicy
  gtff: CopilotForBusinessGeneralPolicy
  o1: CopilotForBusinessPolicy
  o3: CopilotForBusinessPolicy
  o3_mini_ga: boolean
  o_ff: CopilotForBusinessPolicy
  o_fm: CopilotForBusinessGeneralPolicy
  o_f: CopilotForBusinessPolicy
  o_t: CopilotForBusinessGeneralPolicy
  ofo: CopilotForBusinessGeneralPolicy
  mcp: CopilotForBusinessGeneralPolicy
  overages: CopilotForBusinessPolicy
  swe_agent: CopilotForBusinessGeneralPolicy
  spark: CopilotForBusinessGeneralPolicy
  enterprise_name: string | null
  enterprise_slug: string | null
  docsUrls: {
    generalPrivacyStatement: string
  }
}

export type DependentLicense = {
  type: 'dependent'
  name: string
}

export type DirectLicense = {
  type: 'direct'
}

export type FeatureRequest = {
  id: number
  requested_at: string
}

export type FeatureRequestInfo = {
  showFeatureRequest: boolean
  alreadyRequested: boolean
  dismissed: boolean
  featureName: string
  requestPath: string
  isEnterpriseRequest?: boolean
  dismissedAt?: string
  billingEntityId?: string
  latestUsernameRequests: string[]
  amountOfUserRequests: number
}

export type Requester = {
  id: number
  display_login: string
  profile_name: string | null
  requested_at?: string
}

export type UserAssignable = {
  id: number
  avatar_url: string
  email: string
  display_login: string
  org_member: boolean
  profile_name: string | null
  type: typeof SeatType.User
  license?: DependentLicense | DirectLicense
  feature_request?: FeatureRequest
}

export interface UserEmailAssignable extends UserAssignable {
  email_input_csv: boolean
  email: string
}